Restructured "species_details"
NB: This patch really does what a previous one was supposed to do.

The .js can now process all the entries in the "trees" object, which
means that we can any number of trees in the database, and fully maintain
them on the Compara side without having to update the webcode.