Completely rewritten the algorithm to load files for genoverse
For genoverse, the code inside genoverse plugin's dir htdocs/genoverse/vX.X is synced from the genoverse repository without any changes. So in order to override some default genoverse functionality in Ensembl, there's a different folder 'ensembl' inside genoverse plugin's htdocs/genoverse folder. To make sure the code overrides the main genoverse one, each file from the 'ensembl' folder needs to be loaded immidiately after the corresponding core file is loaded. Also if there are further plugins to any of the files, they need to be loaded immidiately after the ones in 'ensembl' folder. To add to the complexity, all files should be ordered according to the order specified in GENOVERSE_FILES_ORDER and the files that are not mentioned in GENOVERSE_FILES_ORDER should be ignored from the 'genoverse' code.